% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absDiff.R
\name{absDiff}
\alias{absDiff}
\title{Returns a vector of the same size as the input vector.
Each value of the new vector is the difference between the current and previous value of the input vector.}
\usage{
absDiff(myVector)
}
\arguments{
\item{data}{input vector}
}
\value{
vector of the same size as the input
}
\description{
Returns a vector of the same size as the input vector.
Each value of the new vector is the difference between the current and previous value of the input vector.
}
\examples{
absDiff(myVector)
}

