% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naFeatures.R
\name{naFeatures}
\alias{naFeatures}
\title{Creates an extra column for every column in the input data that has missing values. 
The newly created column is a factor column (TRUE/FALSE) and represents the existance of NAs or empty strings ("") in the examined column.
The newly created columns are named "na"+name_of_original_column}
\usage{
naFeatures(data)
}
\arguments{
\item{data}{input data}
}
\value{
data frame with the original and the newly created columns
}
\description{
Creates an extra column for every column in the input data that has missing values. 
The newly created column is a factor column (TRUE/FALSE) and represents the existance of NAs or empty strings ("") in the examined column.
The newly created columns are named "na"+name_of_original_column
}
\examples{
naFeatures(myData)
}

