\name{PDMT-package}
\alias{PDMT-package}
\alias{PDMT}
\docType{package}
\title{
Powerful Data Manipulation Tools
}
\description{
Provides functions for automating some useful data manipulation and transformation tasks
}
\details{
\tabular{ll}{
Package: \tab PDMT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-02-24\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Athanasios Tsiaras: <tsiarasa@hotmail.com>
}
\references{

}

\keyword{ package }
\seealso{

~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
naFeatures(myData)
}
